Static cast to UInt8. Explicit float values

+++ src/Protocol/Protocol_1_8.cpp
@@ -1150,7 +1150,7 @@ void cProtocol_1_8_0::SendPlayerMoveLook (Vector3d a_Pos, float a_Yaw, float a_P
 	if (a_IsRelative)
 	{
 		// Set all bits to 1 - makes everything relative
-		Pkt.WriteBEUInt8(-1);
+		Pkt.WriteBEUInt8(static_cast<UInt8>(-1));
 	}
 	else
 	{
